requirements.txt
connected-components-3d
tensorflow==2.3.0
tensorflow-addons
matplotlib
numpy
protobuf
regex
pandas
tqdm
opencv-python
git+https://github.com/yoshihikoueno/dsargparse.git@master
p_tqdm
pyyaml
tensorboard_plugin_profile
ruamel.yaml
